Updates qx
Qx 0.1.0 had a major flaw : The file was named `qx.js` but in the package.json main file was set to `Qx.js` with a capital **Q**. It ran fine on Mac OS X which is case insensitive for this, but it breaks on linux because of this. This solves this problem.